+++ src/vec2.h
@@ -33,12 +33,12 @@ static inline float vec2_dot_product(vec2 a, vec2 b)
 
 static inline float vec2_length(vec2 a)
 {
-    return sqrt(a.x * a.x + a.y * a.y);
+    return sqrtf(a.x * a.x + a.y * a.y);
 }
 
 static inline vec2 vec2_normalize(vec2 a)
 {
-    float scale = sqrt( a.x * a.x + a.y * a.y);
+    float scale = sqrtf( a.x * a.x + a.y * a.y);
     a.x /= scale;
     a.y /= scale;
 
@@ -55,16 +55,16 @@ static inline vec2 vec2_make(float a, float b)
 
 static inline float vec2_angle_degrees(vec2 a)
 {
-    float angle_radians = atan2(a.y, a.x);
-    return (angle_radians / M_PI) * 180.0;
+    float angle_radians = atan2f(a.y, a.x);
+    return (angle_radians / (float)M_PI) * 180.0f;
 }
 
 static inline vec2 vec2_rotate(vec2 a, float b)
 {
-    float theta = b * 180.0 / M_PI;
+    float theta = b * 180.0f / (float)M_PI;
 
-    float cs = cos(theta);
-    float sn = sin(theta);
+    float cs = cosf(theta);
+    float sn = sinf(theta);
 
     vec2 temp;
     temp.x = a.x * cs - a.y * sn;
